pyproject.toml
[project]
name = "upllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "typer>=0.15.2",
    "vastai>=0.2.8",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
upllm = "upllm.cli:app"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "black", "ruff"]
